# resources/schema/swagger/passport.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/login/username/password": {
      "post": {
        "operationId": "LoginController_loginByUsernamePassword",
        "summary": "通过用户名密码登录",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginByUsernamePasswordDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginByUsernamePasswordOKResponseDTO"
                }
              }
            }
          }
        },
        "tags": ["登录"]
      }
    },
    "/login/email/code": {
      "post": {
        "operationId": "LoginController_loginByEmailCode",
        "summary": "通过邮箱验证码登录",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostLoginEmailCodeBodyDTO"
              }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["登录"]
      }
    },
    "/login/phone/code": {
      "post": {
        "operationId": "LoginController_loginByPhoneCode",
        "summary": "通过手机号验证码登录",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostLoginPhoneCodeBodyDTO"
              }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["登录"]
      }
    },
    "/login/qrcode/confirm": {
      "post": {
        "operationId": "LoginController_confirmQrCode",
        "summary": "二维码登录：确认登录",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostLoginQrcodeConfirmBodyDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "type": "number" } }
            }
          }
        },
        "tags": ["登录"],
        "security": [{ "bearer": [] }]
      }
    },
    "/login/qrcode/refresh": {
      "post": {
        "operationId": "LoginController_refreshQrCode",
        "summary": "二维码登录：检查二维码状态并登录",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostLoginQrcodeBodyDTO"
              }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["登录"]
      }
    },
...
  },
  "info": {
    "title": "Passport",
    "description": "#### 一些可以公开的情报\n\n-   本项目所有接口均限流，每个接口每分钟最多只能调用 300 次，超过限制会返回 429 Too Many Requests 错误；\n-   本项目需要鉴权的接口均需要携带 `Authorization` 头，其值为 `Bearer <access_token>`，其中 `<access_token>` 为通过登录接口获取的JWT令牌；\n-   本项目的 JWT 令牌有效期为 30 天，过期后需要重新登录；\n-   本页面支持直接测试已登录接口，点击下面的`Authorize`按钮，输入您的JWT令牌即可直接测试接口（不需要携带`Bearer `）。\n",
    "version": "0.0.0",
    "contact": {
      "name": "Zero",
      "url": "https://naily.cc",
      "email": "gczgroup@qq.com"
    },
    "license": { "name": "UNLICENSED", "url": null }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
      "LoginByUsernamePasswordDTO": {
        "type": "object",
        "properties": {
          "loginType": {
            "type": "string",
            "description": "登录类型",
            "enum": ["WatchOS", "HarmonyOS_Wearable", "Android", "Web", "IOS"]
          },
          "username": { "type": "string", "description": "用户名" },
          "password": { "type": "string", "description": "密码" },
          "loginClient": {
            "type": "string",
            "description": "登录的客户端 用于记录设备/浏览器\n\nWeb端登录时，该字段可以为空"
          },
          "loginDeviceName": {
            "type": "string",
            "description": "登录的设备名\n\nWeb端登录时，该字段可以为空"
          },
          "identifier": {
            "type": "string",
            "description": "登录标识符 整个系统该登录类型的唯一标识符\n\n其中，`Web`端为空则不记录设备"
          }
        },
        "required": ["loginType", "username", "password"]
      },
      "LoginByUsernamePasswordDataUserOKResponseDTO": {
        "type": "object",
        "properties": {
          "userID": { "type": "number", "description": "用户ID", "default": 1 },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "注册时间"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "default": "Zero"
          },
          "ip": { "type": "string", "description": "IP", "default": "0.0.0.0" },
          "saying": {
            "type": "string",
            "description": "个性签名",
            "default": "这个人很懒，什么都没有写哦"
          },
          "avatar": {
            "type": "string",
            "description": "头像地址",
            "default": "https://avatars.githubusercontent.com"
          }
        },
        "required": ["userID", "createdAt", "updatedAt", "username", "ip", "saying", "avatar"]
      },
      "LoginByUsernamePasswordDataOKResponseDTO": {
        "type": "object",
        "properties": {
          "user": {
            "description": "用户信息",
            "allOf": [
              {
                "$ref": "#/components/schemas/LoginByUsernamePasswordDataUserOKResponseDTO"
              }
            ]
          },
          "access_token": { "type": "string", "description": "访问令牌" }
        },
        "required": ["user", "access_token"]
      },
      "LoginByUsernamePasswordOKResponseDTO": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "HTTP状态码",
            "default": 201
          },
          "code": {
            "type": "number",
            "description": "系统状态码",
            "default": 1000
          },
          "message": {
            "type": "string",
            "description": "信息",
            "default": "登录成功"
          },
          "data": {
            "description": "数据",
            "allOf": [
              {
                "$ref": "#/components/schemas/LoginByUsernamePasswordDataOKResponseDTO"
              }
            ]
          }
        },
        "required": ["statusCode", "code", "message", "data"]
      },
      "PostLoginEmailCodeBodyDTO": {
        "type": "object",
        "properties": {
          "loginType": {
            "type": "string",
            "description": "登录设备类型 `请看schema的enum正确传值`",
            "enum": ["WatchOS", "HarmonyOS_Wearable", "Android", "Web", "IOS"]
          },
          "email": { "type": "string", "description": "邮箱" },
          "code": {
            "type": "number",
            "description": "验证码",
            "minimum": 100000,
            "maximum": 999999
          },
          "loginClient": {
            "type": "string",
            "description": "登录的客户端 用于记录设备/浏览器\n\nWeb端登录时，该字段可以为空"
          },
          "loginDeviceName": {
            "type": "string",
            "description": "登录的设备名\n\nWeb端登录时，该字段可以为空"
          },
          "identifier": {
            "type": "string",
            "description": "设备唯一标识符\n\nWeb端登录时，该字段可以为空"
          }
        },
        "required": ["loginType", "email", "code"]
      },
      "PostLoginPhoneCodeBodyDTO": {
        "type": "object",
        "properties": {
          "loginType": {
            "type": "string",
            "description": "登录设备类型 `请看schema的enum正确传值`",
            "enum": ["WatchOS", "HarmonyOS_Wearable", "Android", "Web", "IOS"]
          },
          "phone": { "type": "string", "description": "邮箱" },
          "code": {
            "type": "number",
            "description": "验证码",
            "minimum": 100000,
            "maximum": 999999
          },
          "loginClient": {
            "type": "string",
            "description": "登录的客户端 用于记录设备/浏览器\n\nWeb端登录时，该字段可以为空"
          },
          "loginDeviceName": {
            "type": "string",
            "description": "登录的设备名\n\nWeb端登录时，该字段可以为空"
          },
          "identifier": {
            "type": "string",
            "description": "设备唯一标识符\n\nWeb端登录时，该字段可以为空"
          }
        },
        "required": ["loginType", "phone", "code"]
      },
      "PostLoginQrcodeConfirmBodyDTO": {
        "type": "object",
        "properties": {
          "key": { "type": "string", "description": "二维码key" }
        },
        "required": ["key"]
      },
      "PostLoginQrcodeBodyDTO": {
        "type": "object",
        "properties": {
          "loginType": {
            "type": "string",
            "description": "登录设备类型 `请看schema的enum正确传值`",
            "enum": ["WatchOS", "HarmonyOS_Wearable", "Android", "Web", "IOS"]
          },
          "key": { "type": "string", "description": "二维码key" },
          "loginClient": {
            "type": "string",
            "description": "登录的客户端 用于记录设备/浏览器\n\nWeb端登录时，该字段可以为空"
          },
          "loginDeviceName": {
            "type": "string",
            "description": "登录的设备名\n\nWeb端登录时，该字段可以为空"
          },
          "identifier": {
            "type": "string",
            "description": "设备唯一标识符\n\nWeb端登录时，该字段可以为空"
          }
        },
        "required": ["loginType", "key"]
      },
...
    }
  }
}